Move rewritten folder contents to domains
