+++ client/signup/steps/domains/types.ts
